requirements.txt
fastapi>=0.104.0
fastapi[standard]
uvicorn>=0.23.2
sqlalchemy>=2.0.22
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
loguru>=0.7.2
apscheduler>=3.10.4
boto3>=1.28.62
openai>=1.3.0
google-generativeai>=0.3.1
email-validator>=2.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0